span: create 'substreams' from start to end bounds.
This feature was motivated by glTF .glb files. In them, JSON data is
intermingled with binary data. A glTF parser like `cl-gltf` has a stream
that is the binary data for the whole file, and knows the chunks or
spans of JSON and binary. So we just use json:parse as usual with a new
span/slice/substream of the original stream.

fix not using start as a parameter for make-flexi-stream